+++ changelog.md
@@ -1,8 +1,20 @@
 # Changelog
 
-**./test-parts/possible/valid/route-count.js**
-* Added calls to 'routeResults.checkObject' in the following tests:
-	* "Single"
-	* "Multiple"
-* "Zero" and "Infinity" output numbers instead of objects.
-* 'routeResults.checkObject' is called before the actual route count check.
+**./test-parts/possible/valid/other-cases.js**
+* New file - Remaining unit tests.
+	* Empty Criteria
+	* Impossible Route
+* Adapted from '../../backup'
+
+---
+
+**./test-parts/backup**
+* Moved 'handleArguments' to './possible/valid/other-cases.js' as 'callOtherTestCases'
+* This file is now empty.
+
+---
+
+**./test-parts/test-possible-routes.js**
+* Added requirement for './possible/valid/other-cases' (validOther)
+* Added call to 'validOther'
+* Commented out call to 'validRouteCount'